데일리체크 서버 데이터 변경으로 인한 클라이언트 데이터 변경

+++ UnlimitedTowerClient/TowerMain/UnlimitedTower/Assets/Scripts/Manager/UserDataManager.cs
@@ -32,7 +32,7 @@ public class UserDataManager : MonoSingleton<UserDataManager>
 
     public int currentDayCount = 0;
     public int dayCount = 0;
-    public int currentDayTime = 0;
+    public ulong currentDayTime = 0;
 
     public int usingPartyNum = 1;
